+++ sdk/core/Azure.Core/src/DynamicData/DynamicData.cs
@@ -11,6 +11,7 @@ using System.Reflection;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using Azure.Core.Json;
+using Azure.Core.Serialization;
 
 namespace Azure.Core.Dynamic
 {
@@ -37,43 +38,7 @@ namespace Azure.Core.Dynamic
         {
             _element = element;
             _options = options;
-            _serializerOptions = GetSerializerOptions(options);
-        }
-
-        internal static JsonSerializerOptions GetSerializerOptions(DynamicDataOptions options)
-        {
-            JsonSerializerOptions serializerOptions = new()
-            {
-                Converters =
-                {
-                    new DefaultTimeSpanConverter()
-                }
-            };
-
-            switch (options.CaseMapping)
-            {
-                case DynamicCaseMapping.PascalToCamel:
-                    serializerOptions.PropertyNamingPolicy = JsonNamingPolicy.CamelCase;
-                    break;
-                case DynamicCaseMapping.None:
-                default:
-                    break;
-            }
-
-            switch (options.DateTimeHandling)
-            {
-                case DynamicDateTimeHandling.UnixTime:
-                    serializerOptions.Converters.Add(new UnixTimeDateTimeConverter());
-                    serializerOptions.Converters.Add(new UnixTimeDateTimeOffsetConverter());
-                    break;
-                case DynamicDateTimeHandling.Rfc3339:
-                default:
-                    serializerOptions.Converters.Add(new Rfc3339DateTimeConverter());
-                    serializerOptions.Converters.Add(new Rfc3339DateTimeOffsetConverter());
-                    break;
-            }
-
-            return serializerOptions;
+            _serializerOptions = DynamicDataOptions.ToSerializerOptions(options);
         }
 
         internal void WriteTo(Stream stream)
@@ -101,11 +66,10 @@ namespace Azure.Core.Dynamic
                 return new DynamicData(element, _options);
             }
 
-            // If we're using the PascalToCamel mapping and the strict name lookup
-            // failed, do a second lookup with a camelCase name as well.
-            if (_options.CaseMapping == DynamicCaseMapping.PascalToCamel && char.IsUpper(name[0]))
+            // If the dynamic content has a specified property name format, do a second look-up.
+            if (_options.PropertyNameFormat != JsonPropertyNames.UseExact)
             {
-                if (_element.TryGetProperty(ConvertToCamelCase(name), out element))
+                if (_element.TryGetProperty(FormatPropertyName(name), out element))
                 {
                     if (element.ValueKind == JsonValueKind.Null)
                     {
@@ -120,7 +84,15 @@ namespace Azure.Core.Dynamic
             return null;
         }
 
-        private static string ConvertToCamelCase(string value) => JsonNamingPolicy.CamelCase.ConvertName(value);
+        private string FormatPropertyName(string value)
+        {
+            return _options.PropertyNameFormat switch
+            {
+                JsonPropertyNames.UseExact => value,
+                JsonPropertyNames.CamelCase => JsonNamingPolicy.CamelCase.ConvertName(value),
+                _ => throw new NotSupportedException($"Unknown value for DynamicDataOptions.PropertyNamingConvention: '{_options.PropertyNameFormat}'."),
+            };
+        }
 
         private object? GetViaIndexer(object index)
         {
@@ -172,12 +144,15 @@ namespace Azure.Core.Dynamic
                 value = ConvertType(value);
             }
 
-            if (_options.CaseMapping == DynamicCaseMapping.PascalToCamel)
+            if (_options.PropertyNameFormat == JsonPropertyNames.UseExact ||
+                _element.TryGetProperty(name, out MutableJsonElement _))
             {
-                name = ConvertToCamelCase(name);
+                _element = _element.SetProperty(name, value);
+                return null;
             }
 
-            _element = _element.SetProperty(name, value);
+            // The dynamic content has a specified property name format.
+            _element = _element.SetProperty(FormatPropertyName(name), value);
 
             // Binding machinery expects the call site signature to return an object
             return null;
@@ -372,7 +347,7 @@ namespace Azure.Core.Dynamic
             public override DynamicData Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
             {
                 JsonDocument document = JsonDocument.ParseValue(ref reader);
-                return new DynamicData(new MutableJsonDocument(document, options).RootElement, DynamicDataOptions.Default);
+                return new DynamicData(new MutableJsonDocument(document, options).RootElement, DynamicDataOptions.FromSerializerOptions(options));
             }
 
             public override void Write(Utf8JsonWriter writer, DynamicData value, JsonSerializerOptions options)