+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/src/Internals/AzMonListExtensions.cs
@@ -382,10 +382,5 @@ namespace Azure.Monitor.OpenTelemetry.Exporter.Internals
 
             return "Unknown";
         }
-
-        internal static string? GetAzNameSpace(this AzMonList tagObjects)
-        {
-            return AzMonList.GetTagValue(ref tagObjects, SemanticConventions.AttributeAzureNameSpace)?.ToString();
-        }
     }
 }