+++ sdk/textanalytics/Azure.AI.TextAnalytics/tests/ExtractiveSummarizeTests.cs
@@ -11,7 +11,7 @@ using NUnit.Framework;
 
 namespace Azure.AI.TextAnalytics.Tests
 {
-    [ServiceVersion(Min = TextAnalyticsClientOptions.ServiceVersion.V2022_10_01_Preview)]
+    [ServiceVersion(Min = TextAnalyticsClientOptions.ServiceVersion.V2023_04_01)]
     public class ExtractiveSummarizeTests : TextAnalyticsClientLiveTestBase
     {
         public ExtractiveSummarizeTests(bool isAsync, TextAnalyticsClientOptions.ServiceVersion serviceVersion)
@@ -84,8 +84,7 @@ namespace Azure.AI.TextAnalytics.Tests
         {
             TextAnalyticsClient client = GetClient(useTokenCredential: true);
 
-            ExtractiveSummarizeOperation operation = await client.StartExtractiveSummarizeAsync(s_batchDocuments);
-            await operation.WaitForCompletionAsync();
+            ExtractiveSummarizeOperation operation = await client.ExtractiveSummarizeAsync(WaitUntil.Completed, s_batchDocuments);
             ValidateOperationProperties(operation);
 
             List<ExtractiveSummarizeResultCollection> resultInPages = operation.Value.ToEnumerableAsync().Result;
@@ -93,7 +92,64 @@ namespace Azure.AI.TextAnalytics.Tests
 
             // Take the first page.
             ExtractiveSummarizeResultCollection resultCollection = resultInPages.FirstOrDefault();
-            ValidateSummaryBatchResult(resultCollection, ExtractiveSummarySentencesOrder.Offset);
+            ValidateBatchResult(resultCollection, ExtractiveSummarySentencesOrder.Offset);
+        }
+
+        [RecordedTest]
+        [RetryOnInternalServerError]
+        public async Task ExtractiveSummarizeBatchTest()
+        {
+            TextAnalyticsClient client = GetClient();
+
+            ExtractiveSummarizeOperation operation = await client.ExtractiveSummarizeAsync(WaitUntil.Completed, s_batchDocuments);
+            ValidateOperationProperties(operation);
+
+            List<ExtractiveSummarizeResultCollection> resultInPages = operation.Value.ToEnumerableAsync().Result;
+            Assert.AreEqual(1, resultInPages.Count);
+
+            // Take the first page.
+            ExtractiveSummarizeResultCollection resultCollection = resultInPages.FirstOrDefault();
+            ValidateBatchResult(resultCollection, ExtractiveSummarySentencesOrder.Offset);
+        }
+
+        [RecordedTest]
+        [RetryOnInternalServerError]
+        public async Task ExtractiveSummarizeBatchWaitUntilStartedTest()
+        {
+            TextAnalyticsClient client = GetClient();
+
+            ExtractiveSummarizeOperation operation = await client.ExtractiveSummarizeAsync(WaitUntil.Started, s_batchDocuments);
+            Assert.IsFalse(operation.HasCompleted);
+            Assert.IsFalse(operation.HasValue);
+            Assert.ThrowsAsync<InvalidOperationException>(async () => await Task.Run(() => operation.Value));
+            Assert.ThrowsAsync<InvalidOperationException>(async () => await Task.Run(() => operation.GetValuesAsync()));
+            await operation.WaitForCompletionAsync();
+            Assert.IsTrue(operation.HasCompleted);
+            Assert.IsTrue(operation.HasValue);
+            ValidateOperationProperties(operation);
+        }
+
+        [RecordedTest]
+        [RetryOnInternalServerError]
+        public async Task ExtractiveSummarizeBatchWithStatisticsTest()
+        {
+            TextAnalyticsClient client = GetClient();
+
+            ExtractiveSummarizeOptions options = new ExtractiveSummarizeOptions()
+            {
+                MaxSentenceCount = ExtractiveSummarizeMaxSentenceCount,
+                IncludeStatistics = true,
+            };
+
+            ExtractiveSummarizeOperation operation = await client.ExtractiveSummarizeAsync(WaitUntil.Completed, s_batchDocuments, options);
+            ValidateOperationProperties(operation);
+
+            List<ExtractiveSummarizeResultCollection> resultInPages = operation.Value.ToEnumerableAsync().Result;
+            Assert.AreEqual(1, resultInPages.Count);
+
+            // Take the first page.
+            ExtractiveSummarizeResultCollection resultCollection = resultInPages.FirstOrDefault();
+            ValidateBatchResult(resultCollection, ExtractiveSummarySentencesOrder.Offset, ExtractiveSummarizeMaxSentenceCount, true);
         }
 
         [RecordedTest]
@@ -108,8 +164,7 @@ namespace Azure.AI.TextAnalytics.Tests
                 OrderBy = ExtractiveSummarySentencesOrder.Rank
             };
 
-            ExtractiveSummarizeOperation operation = await client.StartExtractiveSummarizeAsync(s_batchDocuments, options);
-            await operation.WaitForCompletionAsync();
+            ExtractiveSummarizeOperation operation = await client.ExtractiveSummarizeAsync(WaitUntil.Completed, s_batchDocuments, options);
             ValidateOperationProperties(operation);
 
             List<ExtractiveSummarizeResultCollection> resultInPages = operation.Value.ToEnumerableAsync().Result;
@@ -117,7 +172,7 @@ namespace Azure.AI.TextAnalytics.Tests
 
             // Take the first page.
             ExtractiveSummarizeResultCollection resultCollection = resultInPages.FirstOrDefault();
-            ValidateSummaryBatchResult(resultCollection, ExtractiveSummarySentencesOrder.Rank, ExtractiveSummarizeMaxSentenceCount);
+            ValidateBatchResult(resultCollection, ExtractiveSummarySentencesOrder.Rank, ExtractiveSummarizeMaxSentenceCount);
         }
 
         [RecordedTest]
@@ -132,8 +187,7 @@ namespace Azure.AI.TextAnalytics.Tests
                 "",
             };
 
-            ExtractiveSummarizeOperation operation = await client.StartExtractiveSummarizeAsync(documents, "en");
-            await operation.WaitForCompletionAsync();
+            ExtractiveSummarizeOperation operation = await client.ExtractiveSummarizeAsync(WaitUntil.Completed, documents, "en");
             ValidateOperationProperties(operation);
 
             List<ExtractiveSummarizeResultCollection> resultInPages = operation.Value.ToEnumerableAsync().Result;
@@ -152,8 +206,7 @@ namespace Azure.AI.TextAnalytics.Tests
         {
             TextAnalyticsClient client = GetClient();
 
-            ExtractiveSummarizeOperation operation = await client.StartExtractiveSummarizeAsync(s_batchConvenienceDocuments);
-            await operation.WaitForCompletionAsync();
+            ExtractiveSummarizeOperation operation = await client.ExtractiveSummarizeAsync(WaitUntil.Completed, s_batchConvenienceDocuments);
             ValidateOperationProperties(operation);
 
             List<ExtractiveSummarizeResultCollection> resultInPages = operation.Value.ToEnumerableAsync().Result;
@@ -161,7 +214,24 @@ namespace Azure.AI.TextAnalytics.Tests
 
             // Take the first page.
             ExtractiveSummarizeResultCollection resultCollection = resultInPages.FirstOrDefault();
-            ValidateSummaryBatchResult(resultCollection, ExtractiveSummarySentencesOrder.Offset);
+            ValidateBatchResult(resultCollection, ExtractiveSummarySentencesOrder.Offset);
+        }
+
+        [RecordedTest]
+        [RetryOnInternalServerError]
+        public async Task ExtractiveSummarizeBatchConvenienceWaitUntilStartedTest()
+        {
+            TextAnalyticsClient client = GetClient();
+
+            ExtractiveSummarizeOperation operation = await client.ExtractiveSummarizeAsync(WaitUntil.Started, s_batchConvenienceDocuments);
+            Assert.IsFalse(operation.HasCompleted);
+            Assert.IsFalse(operation.HasValue);
+            Assert.ThrowsAsync<InvalidOperationException>(async () => await Task.Run(() => operation.Value));
+            Assert.ThrowsAsync<InvalidOperationException>(async () => await Task.Run(() => operation.GetValuesAsync()));
+            await operation.WaitForCompletionAsync();
+            Assert.IsTrue(operation.HasCompleted);
+            Assert.IsTrue(operation.HasValue);
+            ValidateOperationProperties(operation);
         }
 
         [RecordedTest]
@@ -176,8 +246,7 @@ namespace Azure.AI.TextAnalytics.Tests
                 IncludeStatistics = true,
             };
 
-            ExtractiveSummarizeOperation operation = await client.StartExtractiveSummarizeAsync(s_batchConvenienceDocuments, "en", options);
-            await operation.WaitForCompletionAsync();
+            ExtractiveSummarizeOperation operation = await client.ExtractiveSummarizeAsync(WaitUntil.Completed, s_batchConvenienceDocuments, "en", options);
             ValidateOperationProperties(operation);
 
             List<ExtractiveSummarizeResultCollection> resultInPages = operation.Value.ToEnumerableAsync().Result;
@@ -185,72 +254,11 @@ namespace Azure.AI.TextAnalytics.Tests
 
             // Take the first page.
             ExtractiveSummarizeResultCollection resultCollection = resultInPages.FirstOrDefault();
-            ValidateSummaryBatchResult(resultCollection, ExtractiveSummarySentencesOrder.Offset, ExtractiveSummarizeMaxSentenceCount, true);
+            ValidateBatchResult(resultCollection, ExtractiveSummarySentencesOrder.Offset, ExtractiveSummarizeMaxSentenceCount, true);
         }
 
         [RecordedTest]
         [RetryOnInternalServerError]
-        public async Task ExtractiveSummarizeBatchTest()
-        {
-            TextAnalyticsClient client = GetClient();
-
-            ExtractiveSummarizeOperation operation = await client.StartExtractiveSummarizeAsync(s_batchDocuments);
-            await operation.WaitForCompletionAsync();
-            ValidateOperationProperties(operation);
-
-            List<ExtractiveSummarizeResultCollection> resultInPages = operation.Value.ToEnumerableAsync().Result;
-            Assert.AreEqual(1, resultInPages.Count);
-
-            // Take the first page.
-            ExtractiveSummarizeResultCollection resultCollection = resultInPages.FirstOrDefault();
-            ValidateSummaryBatchResult(resultCollection, ExtractiveSummarySentencesOrder.Offset);
-        }
-
-        [RecordedTest]
-        [RetryOnInternalServerError]
-        public async Task ExtractiveSummarizeBatchWithStatisticsTest()
-        {
-            TextAnalyticsClient client = GetClient();
-
-            ExtractiveSummarizeOptions options = new ExtractiveSummarizeOptions()
-            {
-                MaxSentenceCount = ExtractiveSummarizeMaxSentenceCount,
-                IncludeStatistics = true,
-            };
-
-            ExtractiveSummarizeOperation operation = await client.StartExtractiveSummarizeAsync(s_batchDocuments, options);
-            await operation.WaitForCompletionAsync();
-            ValidateOperationProperties(operation);
-
-            List<ExtractiveSummarizeResultCollection> resultInPages = operation.Value.ToEnumerableAsync().Result;
-            Assert.AreEqual(1, resultInPages.Count);
-
-            // Take the first page.
-            ExtractiveSummarizeResultCollection resultCollection = resultInPages.FirstOrDefault();
-            ValidateSummaryBatchResult(resultCollection, ExtractiveSummarySentencesOrder.Offset, ExtractiveSummarizeMaxSentenceCount, true);
-        }
-
-        [RecordedTest]
-        [RetryOnInternalServerError]
-        public async Task ExtractiveSummarizeBatchConvenienceWithAutoDetectedLanguageTest()
-        {
-            TextAnalyticsClient client = GetClient();
-
-            ExtractiveSummarizeOperation operation = await client.StartExtractiveSummarizeAsync(s_batchConvenienceDocuments, "auto");
-            await operation.WaitForCompletionAsync();
-            ValidateOperationProperties(operation);
-
-            List<ExtractiveSummarizeResultCollection> resultInPages = operation.Value.ToEnumerableAsync().Result;
-            Assert.AreEqual(1, resultInPages.Count);
-
-            // Take the first page.
-            ExtractiveSummarizeResultCollection resultCollection = resultInPages.FirstOrDefault();
-            ValidateSummaryBatchResult(resultCollection, ExtractiveSummarySentencesOrder.Offset, isLanguageAutoDetected: true);
-        }
-
-        [RecordedTest]
-        [RetryOnInternalServerError]
-        [ServiceVersion(Min = TextAnalyticsClientOptions.ServiceVersion.V2022_10_01_Preview)]
         public async Task AnalyzeOperationExtractiveSummarize()
         {
             TextAnalyticsClient client = GetClient();
@@ -260,62 +268,20 @@ namespace Azure.AI.TextAnalytics.Tests
                 DisplayName = "AnalyzeOperationExtractiveSummarize",
             };
 
-            AnalyzeActionsOperation operation = await client.StartAnalyzeActionsAsync(s_batchConvenienceDocuments, batchActions);
-            await operation.WaitForCompletionAsync();
+            AnalyzeActionsOperation operation = await client.AnalyzeActionsAsync(WaitUntil.Completed, s_batchConvenienceDocuments, batchActions);
+            Assert.IsTrue(operation.HasCompleted);
 
-            AnalyzeActionsResult resultCollection = operation.Value.ToEnumerableAsync().Result.FirstOrDefault();
-            IReadOnlyCollection<ExtractiveSummarizeActionResult> extractiveSummarizeActionsResults = resultCollection.ExtractiveSummarizeResults;
-            Assert.IsNotNull(extractiveSummarizeActionsResults);
+            AnalyzeActionsResult actionsResult = operation.Value.ToEnumerableAsync().Result.FirstOrDefault();
+            IReadOnlyCollection<ExtractiveSummarizeActionResult> extractiveSummarizeActionResults = actionsResult.ExtractiveSummarizeResults;
+            Assert.IsNotNull(extractiveSummarizeActionResults);
 
-            ExtractiveSummarizeResultCollection extractiveSummarizeDocumentsResults = extractiveSummarizeActionsResults.FirstOrDefault().DocumentsResults;
-            ValidateSummaryBatchResult(extractiveSummarizeDocumentsResults, ExtractiveSummarySentencesOrder.Offset);
-        }
-
-        [RecordedTest]
-        [RetryOnInternalServerError]
-        [ServiceVersion(Max = TextAnalyticsClientOptions.ServiceVersion.V2022_05_01)]
-        public void AnalyzeOperationExtractiveSummarizeActionNotSupported()
-        {
-            TestDiagnostics = false;
-            TextAnalyticsClient client = GetClient();
-            TextAnalyticsActions batchActions = new()
-            {
-                ExtractiveSummarizeActions = new[]
-                {
-                    new ExtractiveSummarizeAction(),
-                },
-            };
-
-            NotSupportedException ex = Assert.ThrowsAsync<NotSupportedException>(async () => await client.StartAnalyzeActionsAsync(s_batchDocuments, batchActions));
-            Assert.That(ex.Message.EndsWith("Use service API version 2022-10-01-preview or newer."));
-        }
-
-        [RecordedTest]
-        [RetryOnInternalServerError]
-        public async Task AnalyzeOperationExtractiveSummarizeWithAutoDetectedLanguageTest()
-        {
-            TextAnalyticsClient client = GetClient();
-            List<string> documents = s_batchConvenienceDocuments;
-            TextAnalyticsActions actions = new()
-            {
-                ExtractiveSummarizeActions = new List<ExtractiveSummarizeAction>() { new ExtractiveSummarizeAction() },
-                DisplayName = "ExtractiveSummarizeWithAutoDetectedLanguage",
-            };
-
-            AnalyzeActionsOperation operation = await client.StartAnalyzeActionsAsync(documents, actions, "auto");
-            await operation.WaitForCompletionAsync();
-
-            // Take the first page.
-            AnalyzeActionsResult resultCollection = operation.Value.ToEnumerableAsync().Result.FirstOrDefault();
-            IReadOnlyCollection<ExtractiveSummarizeActionResult> actionResults = resultCollection.ExtractiveSummarizeResults;
-            Assert.IsNotNull(actionResults);
-
-            ExtractiveSummarizeResultCollection results = actionResults.FirstOrDefault().DocumentsResults;
-            ValidateSummaryBatchResult(results, ExtractiveSummarySentencesOrder.Offset, isLanguageAutoDetected: true);
+            ExtractiveSummarizeResultCollection resultCollection = extractiveSummarizeActionResults.FirstOrDefault().DocumentsResults;
+            ValidateBatchResult(resultCollection, ExtractiveSummarySentencesOrder.Offset);
         }
 
         private void ValidateOperationProperties(ExtractiveSummarizeOperation operation)
         {
+            Assert.IsTrue(operation.HasCompleted);
             Assert.AreNotEqual(new DateTimeOffset(), operation.CreatedOn);
             // TODO: Re-enable this check (https://github.com/Azure/azure-sdk-for-net/issues/31855).
             // Assert.AreNotEqual(new DateTimeOffset(), operation.LastModified);
@@ -326,7 +292,7 @@ namespace Azure.AI.TextAnalytics.Tests
             }
         }
 
-        private void ValidateSummaryDocumentResult(
+        private void ValidateDocumentResult(
             IReadOnlyCollection<ExtractiveSummarySentence> sentences,
             int maxSentenceCount,
             ExtractiveSummarySentencesOrder expectedOrder)
@@ -361,12 +327,11 @@ namespace Azure.AI.TextAnalytics.Tests
             }
         }
 
-        private void ValidateSummaryBatchResult(
+        private void ValidateBatchResult(
             ExtractiveSummarizeResultCollection results,
             ExtractiveSummarySentencesOrder expectedOrder,
             int maxSentenceCount = DefaultSummaryMaxSentenceCount,
-            bool includeStatistics = default,
-            bool isLanguageAutoDetected = default)
+            bool includeStatistics = default)
         {
             Assert.That(results.ModelVersion, Is.Not.Null.And.Not.Empty);
 
@@ -400,22 +365,7 @@ namespace Azure.AI.TextAnalytics.Tests
                     Assert.AreEqual(0, result.Statistics.TransactionCount);
                 }
 
-                if (isLanguageAutoDetected)
-                {
-                    Assert.IsNotNull(result.DetectedLanguage);
-                    Assert.That(result.DetectedLanguage.Value.Name, Is.Not.Null.And.Not.Empty);
-                    Assert.That(result.DetectedLanguage.Value.Iso6391Name, Is.Not.Null.And.Not.Empty);
-                    Assert.GreaterOrEqual(result.DetectedLanguage.Value.ConfidenceScore, 0.0);
-                    Assert.LessOrEqual(result.DetectedLanguage.Value.ConfidenceScore, 1.0);
-                    Assert.IsNotNull(result.DetectedLanguage.Value.Warnings);
-                    Assert.IsEmpty(result.DetectedLanguage.Value.Warnings);
-                }
-                else
-                {
-                    Assert.IsNull(result.DetectedLanguage);
-                }
-
-                ValidateSummaryDocumentResult(result.Sentences, maxSentenceCount, expectedOrder);
+                ValidateDocumentResult(result.Sentences, maxSentenceCount, expectedOrder);
             }
         }
     }