+++ sdk/textanalytics/Azure.AI.TextAnalytics/tests/samples/Sample12_AbstractiveSummarizeAsync.cs
@@ -72,8 +72,7 @@ namespace Azure.AI.TextAnalytics.Samples
             };
 
             // Perform the text analysis operation.
-            AbstractiveSummarizeOperation operation = client.StartAbstractiveSummarize(batchedDocuments, options);
-            await operation.WaitForCompletionAsync();
+            AbstractiveSummarizeOperation operation = client.AbstractiveSummarize(WaitUntil.Completed, batchedDocuments, options);
 
             Console.WriteLine($"The operation has completed.");
             Console.WriteLine();